rename: 기획 변경에 따른 폴더 이동
- 문제집 에디터 페이지가 모달로 변경됨
